works and is still opens dot menu fast
